Floats converted to double

+++ QLepp2DGui/openglwidget.h
@@ -81,7 +81,7 @@ private:
     float m_yCamPos;
     float m_zCamPos;
 
-    QVector<GLfloat> m_data;
+    QVector<GLdouble> m_data;
     bool m_dataAlreadyLoaded;
 
     Model *m_model;
